requirements.txt
# Crypto Futures AI Trading Agent - Dependencies

# Exchange connectivity
ccxt>=4.0.0                # Unified exchange API (Binance, Bybit, Bitunix, etc.)

# Data & Analysis
pandas>=2.0.0              # DataFrames for candle data
numpy>=1.24.0              # Numerical computations (TA indicators computed in pure pandas)

# AI Models
groq>=1.0.0                # Groq API (Llama 3.3 70B - FREE)

# HTTP & Data
requests>=2.31.0           # HTTP client for news APIs

# Telegram signal scraping
telethon>=1.35.0           # Telegram client for channel scraping

# AI Vision (chart analysis)
google-genai>=1.0.0        # Google Gemini API (chart image analysis - FREE)

# Configuration
python-dotenv>=1.0.0       # .env file loading
